Update according to the new .urcheon dir name

+++ unv-dpk.nix
@@ -26,7 +26,8 @@ stdenv.mkDerivation rec {
     cp -v --no-preserve=mode ${unvanquished-vms}/* .
     rm *.so
     rm .git* -r
-    rm .pakinfo -r
+    [ -d .pakinfo ] && rm .pakinfo -r
+    [ -d .urcheon ] && rm .urcheon -r
     7z -tzip -mx=9 a ../${filename}_${version}.dpk .
     cd ..
   '';
